Refact: Renaming main activity to calypso activity
