+++ java/example/calypso/android/nfc/src/main/AndroidManifest.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="org.eclipse.keyple.example.calypso.android.nfc">
-
-    <!-- Required for NFC-->
+    <!-- Required for NFC -->
     <uses-permission android:name="android.permission.NFC" />
+
     <uses-feature
         android:name="android.hardware.nfc"
         android:required="true" />
@@ -16,22 +16,22 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-
-
+        <activity android:name="org.eclipse.keyple.example.activity.AbstractExampleActivity"></activity>
         <activity
-            android:name="org.eclipse.keyple.example.activity.MainActivity"
+            android:name="org.eclipse.keyple.example.activity.CalypsoExamplesActivity"
             android:label="Keyple Android NFC"
             android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-            <!-- Required for NFC KEYPLE-->
+            <!-- Required for NFC KEYPLE -->
             <intent-filter>
                 <action android:name="android.nfc.action.TECH_DISCOVERED" />
             </intent-filter>
 
-            <!-- Required for NFC KEYPLE-->
+            <!-- Required for NFC KEYPLE -->
             <meta-data
                 android:name="android.nfc.action.TECH_DISCOVERED"
                 android:resource="@xml/tech_list" />
