+++ java/example/calypso/android/nfc/src/main/kotlin/org/eclipse/keyple/example/activity/AbstractExampleActivity.kt
@@ -1,0 +1,132 @@
+package org.eclipse.keyple.example.activity
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.view.WindowManager
+import androidx.appcompat.app.ActionBarDrawerToggle
+import androidx.appcompat.app.AlertDialog
+import androidx.appcompat.widget.Toolbar
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import com.google.android.material.navigation.NavigationView
+import kotlinx.android.synthetic.main.activity_calypso_examples.drawerLayout
+import kotlinx.android.synthetic.main.activity_calypso_examples.eventRecyclerView
+import kotlinx.android.synthetic.main.activity_calypso_examples.navigationView
+import kotlinx.android.synthetic.main.activity_calypso_examples.toolbar
+import org.eclipse.keyple.core.seproxy.SeProxyService
+import org.eclipse.keyple.core.seproxy.SeReader
+import org.eclipse.keyple.core.seproxy.event.ObservableReader
+import org.eclipse.keyple.core.seproxy.protocol.SeCommonProtocols
+import org.eclipse.keyple.example.adapter.EventAdapter
+import org.eclipse.keyple.example.calypso.android.nfc.R
+import org.eclipse.keyple.example.model.ChoiceEventModel
+import org.eclipse.keyple.example.model.EventModel
+import org.eclipse.keyple.example.util.configFlags
+import org.eclipse.keyple.example.util.configProtocol
+import org.eclipse.keyple.plugin.android.nfc.AndroidNfcPluginFactory
+import org.eclipse.keyple.plugin.android.nfc.AndroidNfcReader
+import timber.log.Timber
+import java.util.SortedSet
+
+abstract class AbstractExampleActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelectedListener, ObservableReader.ReaderObserver {
+
+    /**
+     * Variables for event window
+     */
+    protected lateinit var readers: SortedSet<SeReader>
+    private lateinit var adapter: RecyclerView.Adapter<*>
+    private lateinit var layoutManager: RecyclerView.LayoutManager
+    protected val events = arrayListOf<EventModel>()
+
+    protected lateinit var reader: AndroidNfcReader
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        initContentView()
+
+        /**
+         * Init recycler view
+         */
+        adapter = EventAdapter(events)
+        layoutManager = LinearLayoutManager(this)
+        eventRecyclerView.layoutManager = layoutManager
+        eventRecyclerView.adapter = adapter
+
+        /**
+         * Init menu
+         */
+        navigationView.setNavigationItemSelectedListener(this)
+        val toggle = ActionBarDrawerToggle(this, drawerLayout, toolbar, R.string.open_navigation_drawer, R.string.close_navigation_drawer)
+        drawerLayout.addDrawerListener(toggle)
+        toggle.syncState()
+
+
+        /**
+         * Register AndroidNfc plugin
+         */
+        SeProxyService.getInstance().registerPlugin(AndroidNfcPluginFactory())
+
+        /**
+         *  remove the observer if it already exist
+         */
+        reader = SeProxyService.getInstance().plugins.first().readers.first() as AndroidNfcReader
+        reader.configFlags(presenceCheckDelay = 100, noPlateformSound = 0, skipNdefCheck = 0)
+
+        (reader as ObservableReader).addObserver(this)
+
+        // with this protocol settings we activate the nfc for ISO1443_4 protocol
+        reader.configProtocol(SeCommonProtocols.PROTOCOL_ISO14443_4)
+
+        window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
+    }
+
+    protected fun initActionBar(toolbar: Toolbar, title: String, subtitle: String) {
+        setSupportActionBar(toolbar)
+        val actionBar = supportActionBar
+        actionBar?.title = title
+        actionBar?.subtitle = subtitle
+    }
+
+    protected fun showAlertDialog(t: Throwable) {
+        val builder = AlertDialog.Builder(this@AbstractExampleActivity)
+        builder.setTitle(R.string.alert_dialog_title)
+        builder.setMessage(getString(R.string.alert_dialog_message, t.message))
+        val dialog = builder.create()
+        dialog.show()
+    }
+
+    protected fun initFromBackgroundTextView() {
+        addResultEvent("Smartcard detected while in background...")
+    }
+
+    protected fun clearEvents() {
+        events.clear()
+        adapter.notifyDataSetChanged()
+    }
+
+    protected fun addHeaderEvent(message: String) {
+        events.add(EventModel(EventModel.TYPE_HEADER, message))
+        adapter.notifyItemInserted(events.lastIndex)
+        Timber.d("Header: %s", message)
+    }
+
+    protected fun addActionEvent(message: String) {
+        events.add(EventModel(EventModel.TYPE_ACTION, message))
+        adapter.notifyItemInserted(events.lastIndex)
+        Timber.d("Action: %s", message)
+    }
+
+    protected fun addResultEvent(message: String) {
+        events.add(EventModel(EventModel.TYPE_RESULT, message))
+        adapter.notifyItemInserted(events.lastIndex)
+        Timber.d("Result: %s", message)
+    }
+
+    protected fun addChoiceEvent(title: String, choices: List<String>, callback: (choice: String) -> Unit) {
+        events.add(ChoiceEventModel(title, choices, callback))
+        adapter.notifyItemInserted(events.lastIndex)
+        Timber.d("Choice: %s: %s", title, choices.toString())
+    }
+
+    abstract fun initContentView()
+}
