+++ java/example/calypso/android/nfc/src/main/res/layout/activity_calypso_examples.xml
@@ -8,12 +8,12 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     tools:openDrawer="start"
-    tools:context="org.eclipse.keyple.example.activity.MainActivity">
+    tools:context="org.eclipse.keyple.example.activity.CalypsoExamplesActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="org.eclipse.keyple.example.activity.MainActivity">
+        tools:context="org.eclipse.keyple.example.activity.CalypsoExamplesActivity">
 
 
         <androidx.appcompat.widget.Toolbar
@@ -48,7 +48,7 @@
         android:layout_gravity="start"
         android:fitsSystemWindows="true"
         app:itemTextColor="@color/colorAccent"
-        app:menu="@menu/main_menu"
+        app:menu="@menu/calypso_examples_menu"
         app:headerLayout="@layout/nav_drawer_header" />
 
 </androidx.drawerlayout.widget.DrawerLayout>
