+++ java/example/calypso/android/nfc/src/main/res/menu/calypso_examples_menu.xml
